feat : devmode 모듈 분리
