+++ data/src/main/kotlin/com/wepli/data/di/repository/RepositoryModule.kt
@@ -17,8 +17,6 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import com.wepli.domain.devmode.apilog.repository.DebugApiLogRepository
 import com.wepli.data.devmode.apilog.repository.DebugApiLogRepositoryImpl
-import com.wepli.domain.devmode.fcm.repository.DevModeFcmRepository
-import com.wepli.data.devmode.fcm.repository.DevModeFcmRepositoryImpl
 import repository.setting.SettingRepository
 import com.wepli.data.setting.repository.SettingRepositoryImpl
 import repository.youtube.YoutubeRepository
@@ -84,8 +82,4 @@ interface RepositoryModule {
     @Binds
     @Singleton
     fun bindDebugApiLogRepository(debugApiLogRepositoryImpl: DebugApiLogRepositoryImpl): DebugApiLogRepository
-
-    @Binds
-    @Singleton
-    fun bindDevFcmRepository(devModeFcmRepositoryImpl: DevModeFcmRepositoryImpl): DevModeFcmRepository
 }
